글쓰기에서 익명 게시판이 아니면, 전 화면으로 돌아가게 수정 관리자 페이지에 전체 문의 조회/답변 기능 추가 1:1 문의 접수 페이지 (/center/create & /mypage/service/asking) 추가 마이페이지에서 1:1문의내역 페이지 구현

+++ src/components/pages/community/Write.js
@@ -9,6 +9,7 @@ import {createBoardPosts, image} from "../../http/tming";
 import {errMsg} from "../../http/util";
 import Optional from "../../primitive/Optional/Optional";
 import Wysiwyg from "../../primitive/WYSIWYG/WYSIWYG";
+import {authorized} from "../../utils/utils";
 
 
 class Write extends Component {
@@ -18,13 +19,22 @@ class Write extends Component {
   };
 
   componentDidMount() {
-    const {uiKit}= this.props;
+    const {uiKit, auth, location, history}= this.props;
+    const {category}= urlQuery(location);
+    if(!authorized(auth) && category!== 'anonymous'){
+      console.log(auth, category);
+      uiKit.toaster.cooking('익명 게시판이 아니면 로그인이 필요합니다');
+      history.goBack();
+      return;
+    }
+
     uiKit.destroyAll();
     this.unblock= this.props.history.block('작성한 글이 모두 사라집니다');
   }
 
   componentWillUnmount() {
-    this.unblock();
+    if(this.unblock)
+      this.unblock();
   }
 
   createBoardPost= async ()=>{
